Split detectors.py into individual files
These were buildpacks now anyway, not detectors.
