+++ repo2docker/buildpacks/legacy.py
@@ -1,0 +1,60 @@
+"""
+Generates a variety of Dockerfiles based on an input matrix
+"""
+import textwrap
+from traitlets.config import LoggingConfigurable
+from traitlets import Unicode, Set, List, Dict, Tuple, default
+from textwrap import dedent
+import jinja2
+import tarfile
+import io
+import os
+import stat
+import re
+import docker
+from .docker import DockerBuildPack
+
+class LegacyBinderDockerBuildPack(DockerBuildPack):
+
+    name = 'Legacy Binder Dockerfile'
+    dockerfile = '._binder.Dockerfile'
+
+    dockerfile_appendix = Unicode(dedent(r"""
+    USER root
+    COPY . /home/main/notebooks
+    RUN chown -R main:main /home/main/notebooks
+    USER main
+    WORKDIR /home/main/notebooks
+    ENV PATH /home/main/anaconda2/envs/python3/bin:$PATH
+    RUN conda install -yq -n python3 notebook==5.0.0 ipykernel==4.6.0 && \
+        conda remove -yq -n python3 nb_conda_kernels && \
+        conda install -yq -n root ipykernel==4.6.0 && \
+        /home/main/anaconda2/envs/python3/bin/ipython kernel install --sys-prefix && \
+        /home/main/anaconda2/bin/ipython kernel install --prefix=/home/main/anaconda2/envs/python3 && \
+        /home/main/anaconda2/bin/ipython kernel install --sys-prefix
+    ENV JUPYTER_PATH /home/main/anaconda2/share/jupyter:$JUPYTER_PATH
+    CMD jupyter notebook --ip 0.0.0.0
+    """), config=True)
+
+    def render(self):
+        with open('Dockerfile') as f:
+            return f.read() + self.dockerfile_appendix
+
+    def build(self, image_spec):
+        with open(self.dockerfile, 'w') as f:
+            f.write(self.render())
+        return super().build(image_spec)
+
+    def detect(self):
+        try:
+            with open('Dockerfile', 'r') as f:
+                for line in f:
+                    if line.startswith('FROM'):
+                        if 'andrewosh/binder-base' in line.split('#')[0].lower():
+                            return True
+                        else:
+                            return False
+        except FileNotFoundError:
+            pass
+
+        return False
